feat(PPDSC-2020): update snapshots
