+++ src/radio-button/index.tsx
@@ -1,3 +1,3 @@
 export {RadioButton} from './radio-button';
-export type {RadioButtonProps} from './types';
+export type {RadioButtonProps, RadioGroupProps} from './types';
 export {RadioGroup} from './radio-group';
