+++ src/radio-button/types.ts
@@ -4,5 +4,19 @@ export type RadioButtonIconProps = SwitchBaseIconProps;
 
 export type RadioButtonProps = Omit<
   SwitchBaseProps,
-  'path' | 'defaultSwitchComponent' | 'type'
+  'path' | 'defaultSwitchComponent' | 'type' | 'defaultChecked' | 'defaultValue'
 >;
+
+export type RadioGroupContextValue = {
+  name: string | undefined;
+  onChange: (event: React.ChangeEvent<HTMLInputElement>, value: string) => void;
+  value: string | undefined;
+};
+
+export type RadioGroupProps = {
+  defaultValue?: string;
+  value?: string;
+  onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  name?: string;
+  children?: React.ReactNode;
+};
